requirements.txt
discord
urllib3
python-dotenv
requests